requirements.txt
# hikari
hikari[speedups]

hikari-tanjun
hikari-tanchi
alluka

# meta
nox
typing_extensions
PyYAML
psutil
python-dotenv

black
isort
pyright
mypy

# db
asyncpg
asyncpg-stubs-h1

# utils
thefuzz[speedup]
attrs
pydantic

pysaucenao
genshin[geetest]

# debugging
devtools

# web
fastapi
uvicorn[standard]
